# Validate the metric names and labels in the remote write handler