@roidelapluie 2023-04-05
there is interest in this from the community, could you rebase so we know it works with label string? thanks